Marketing button - Class migration.
