+++ classes/class-uagb-block-helper.php
@@ -4476,12 +4476,13 @@ if ( ! class_exists( 'UAGB_Block_Helper' ) ) {
 			);
 
 			// @codingStandardsIgnoreEnd
+			$base_selector = ( $attr['classMigrate'] ) ? '.uagb-block-' : '#uagb-marketing-btn-';
 
-			$desktop = UAGB_Helper::generate_css( $selectors, '#uagb-marketing-btn-' . $id );
+			$desktop = UAGB_Helper::generate_css( $selectors, $base_selector . $id );
 
-			$tablet = UAGB_Helper::generate_css( $t_selectors, '#uagb-marketing-btn-' . $id );
+			$tablet = UAGB_Helper::generate_css( $t_selectors, $base_selector . $id );
 
-			$mobile = UAGB_Helper::generate_css( $m_selectors, '#uagb-marketing-btn-' . $id );
+			$mobile = UAGB_Helper::generate_css( $m_selectors, $base_selector . $id );
 
 			$generated_css = array(
 				'desktop' => $desktop,
