+++ src/blocks/marketing-button/attributes.js
@@ -4,6 +4,10 @@ const attributes = {
 	block_id: {
 		type: "string"
 	},
+	classMigrate: {
+		type: "boolean",
+		default: false
+	},
 	align: {
 		type: "string",
 		default: "center"
